+++ app/api/CardanoClientApi.js
@@ -101,6 +101,29 @@ const unsetTermsOfUseAcceptanceFromLocalStorage = () => new Promise((resolve) =>
   });
 });
 
+const getSendLogsChoiceFromLocalStorage = () => new Promise((resolve, reject) => {
+  localStorage.get('sendLogsChoice', (error, response) => {
+    if (error) return reject(error);
+    if (typeof response.sendLogs === 'undefined') {
+      return resolve(null);
+    }
+    resolve(response.sendLogs);
+  });
+});
+
+const setSendLogsChoiceInLocalStorage = (sendLogs) => new Promise((resolve, reject) => {
+  localStorage.set('sendLogsChoice', { sendLogs }, (error) => {
+    if (error) return reject(error);
+    resolve();
+  });
+});
+
+const unsetSendLogsChoiceFromLocalStorage = () => new Promise((resolve) => {
+  localStorage.remove('sendLogsChoice', () => {
+    resolve();
+  });
+});
+
 export default class CardanoClientApi {
 
   notifyCallbacks = [];
@@ -367,23 +390,6 @@ export default class CardanoClientApi {
     }
   }
 
-  // PRIVATE
-
-  _onNotify = (rawMessage: string) => {
-    Log.debug('CardanoClientApi::notify message: ', rawMessage);
-    // TODO: "ConnectionClosed" messages are not JSON parsable … so we need to catch that case here!
-    let message = rawMessage;
-    if (message !== 'ConnectionClosed') {
-      message = JSON.parse(rawMessage);
-    }
-    this.notifyCallbacks.forEach(cb => cb.message(message));
-  };
-
-  _onNotifyError = (error: Error) => {
-    Log.error('CardanoClientApi::notify error: ' + stringifyError(error));
-    this.notifyCallbacks.forEach(cb => cb.error(error));
-  };
-
   async nextUpdate() {
     Log.debug('CardanoClientApi::nextUpdate called');
     let nextUpdate = null;
@@ -494,6 +500,30 @@ export default class CardanoClientApi {
     }
   }
 
+  async setSendLogsChoice(sendLogs: boolean) {
+    Log.debug('CardanoClientApi::setSendLogsChoice called: ', sendLogs);
+    try {
+      await setSendLogsChoiceInLocalStorage(sendLogs);
+      Log.debug('CardanoClientApi::setSendLogsChoice success: ', sendLogs);
+      return sendLogs;
+    } catch (error) {
+      Log.error('CardanoClientApi::setSendLogsChoice error: ' + stringifyError(error));
+      throw new GenericApiError();
+    }
+  }
+
+  async getSendLogsChoice() {
+    Log.debug('CardanoClientApi::getSendLogsChoice called');
+    try {
+      const logs = await getSendLogsChoiceFromLocalStorage();
+      Log.debug('CardanoClientApi::getSendLogsChoice success: ', logs);
+      return logs;
+    } catch (error) {
+      Log.error('CardanoClientApi::getSendLogsChoice error: ' + stringifyError(error));
+      throw new GenericApiError();
+    }
+  }
+
   async getTermsOfUseAcceptance() {
     Log.debug('CardanoClientApi::getTermsOfUseAcceptance called');
     try {
@@ -539,6 +569,7 @@ export default class CardanoClientApi {
     Log.debug('CardanoClientApi::testReset called');
     await unsetUserLocaleFromLocalStorage(); // TODO: remove after saving locale to API is restored
     await unsetTermsOfUseAcceptanceFromLocalStorage();
+    await unsetSendLogsChoiceFromLocalStorage();
     try {
       const response = await ClientApi.testReset();
       Log.debug('CardanoClientApi::testReset success: ', stringifyData(response));